i18n(ko): make Korean selectable and enforce locale registration parity
The Korean locale landed registered with i18next but not wired into the
places that actually surface it, so nobody could select it.

- Add Korean to the Settings language picker. Until now a user whose browser
  locale was Korean got a Korean UI while the selector read "English", and no
  user could choose Korean at all.
- Add a Korean form to the service-worker string table, so notifications no
  longer fall back to English.
- Use `displayCount` in the four unread/new-message templates. `main` renamed
  the placeholder after the locale was written, which bypassed the saturation
  cap: Korean showed "1000" where every other locale shows "999+".
- Update the language count in the README.

Nothing checked that the picker and `supportedLanguages` agree, which is why
green CI shipped an unselectable locale. The locale registry now lives in a
side-effect-free `i18n/languages.ts`, and the i18n suite asserts that the
picker, `supportedLanguages`, the shipped locale files, and the service-worker
table all cover the same set, in both directions.

Also polish a few Korean strings: first-person `나` for the current user in
reactor lists, standard `라이선스`, `닉네임` for consistency with the room and
command labels, distinct labels for three different "more" actions, one folded
sentence in the moderated-room hint, and command usage metavariables back to
English to match the other 33 locales.
